refactor: consolidate all the methods in a single class and apply suggestions from the feedback
